requirement.txt
torch=2.6.0+cu118
numpy=2.2.3
pillow=11.1.0
opencv-python=4.11.0.86